Fix path to child theme
BUGFIX: Original code was `get_template_directory()` when `is_child_theme()` returns `TRUE`. This means we would return the "parent" theme directory when using a child theme. 

However, in two projects so far where I have implemented this code with a child theme, my natural expectation is that when using a child theme, I would expect `/htmx-templates/` to be found in the child theme.